Split out code for threads to different files
